Avoid potential overflow in WndProc

+++ EDDiscovery/EDDiscoveryForm.cs
@@ -1064,7 +1064,7 @@ namespace EDDiscovery
         protected override void WndProc(ref Message m)
         {
             // Compatibility movement for Mono
-            if (m.Msg == WM.LBUTTONDOWN && (int)m.WParam == 1 && !theme.WindowsFrame)
+            if (m.Msg == WM.LBUTTONDOWN && m.WParam == (IntPtr)1 && !theme.WindowsFrame)
             {
                 int x = unchecked((short)((uint)m.LParam & 0xFFFF));
                 int y = unchecked((short)((uint)m.LParam >> 16));
@@ -1074,7 +1074,7 @@ namespace EDDiscovery
                 m.Result = IntPtr.Zero;
                 this.Capture = true;
             }
-            else if (m.Msg == WM.MOUSEMOVE && (int)m.WParam == 1 && _window_dragging)
+            else if (m.Msg == WM.MOUSEMOVE && m.WParam == (IntPtr)1 && _window_dragging)
             {
                 int x = unchecked((short)((uint)m.LParam & 0xFFFF));
                 int y = unchecked((short)((uint)m.LParam >> 16));
@@ -1098,7 +1098,7 @@ namespace EDDiscovery
                 base.WndProc(ref m);
                 //System.Diagnostics.Debug.WriteLine( Environment.TickCount + " Res " + ((int)m.Result));
 
-                if ((int)m.Result == HT.CLIENT)
+                if (m.Result == (IntPtr)HT.CLIENT)
                 {
                     int x = unchecked((short)((uint)m.LParam & 0xFFFF));
                     int y = unchecked((short)((uint)m.LParam >> 16));
